Update ReentrantLock usage to use try-with-resource model
The ServiceUse.lock() method returns an AutoCloseable so the
caller can use try-with-resources to obtain and then close the lock.

This is much closer to intrinsic locking in coding practice.

Instead of using a long-ish timeout for deadlock detection, which
could be wrong on slow systems, we now use a map to hold the thread
to lock parings and examine the lock chain to see if the current
thread is in the chain.

Signed-off-by: Hannes Wellmann <wellmann.hannes1@gmx.net>
Signed-off-by: BJ Hargrave <hargrave@us.ibm.com>